+++ Assets/Scripts/Shared Scripts/IDamage.cs
@@ -1,9 +1,8 @@
-//Worked on by : Jacob Irvin, Joshua Furber
-
+// Worked on by - Jacob Irvin, Joshua Furber
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
 public interface IDamage {
-    public void TakeDamage(int damage);
+    public void TakeDamage(float damage);
 }
